requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
google-genai==1.31.0
pydantic==2.11.7
